делит код в main.js, на отдельные ES2015-модули
